ref(rename): Remove suffix v2 from stack trace components
