+++ static/app/components/events/interfaces/deprecatedThreads/content.tsx
