+++ static/app/components/events/interfaces/deprecatedThreads/index.tsx
