+++ static/app/components/events/interfaces/deprecatedThreads/threadSelector/getRelevantFrame.tsx
