+++ static/app/components/events/interfaces/deprecatedThreads/threadSelector/getThreadException.tsx
