+++ static/app/components/events/interfaces/deprecatedThreads/threadSelector/getThreadStacktrace.tsx
