+++ static/app/components/events/interfaces/deprecatedThreads/threadSelector/header.tsx
