+++ static/app/components/events/interfaces/deprecatedThreads/threadSelector/index.tsx
