+++ static/app/components/events/interfaces/deprecatedThreads/threadSelector/option.tsx
