+++ static/app/components/events/interfaces/deprecatedThreads/threadSelector/selectedOption.tsx
