+++ static/app/components/events/interfaces/deprecatedThreads/threadSelector/styles.tsx
